Add find a course unit test and rename test file
I added a test that verifies that the find a course function works, it also verifies that if no course exists by that id it returns None

issue #1
